API doc

+++ app/html/src/Controller/UserController.php
@@ -9,38 +9,36 @@ use App\Repository\CustomerRepository;
 use Doctrine\ORM\EntityManagerInterface;
 use Exception;
 use JMS\Serializer\SerializationContext;
-use JMS\Serializer\Serializer;
-use Symfony\Contracts\Cache\ItemInterface;
+use JMS\Serializer\SerializerInterface as JMSInterface;
 use Knp\Component\Pager\PaginatorInterface;
+use OpenApi\Annotations as OA;
+use Symfony\Contracts\Cache\ItemInterface;
 use Symfony\Contracts\Cache\CacheInterface;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
-use Symfony\Component\HttpFoundation\JsonResponse;
-use Symfony\Component\Serializer\SerializerInterface;
 use Symfony\Component\Validator\Validator\ValidatorInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
-use Symfony\Component\Serializer\Exception\NotEncodableValueException;
 use Symfony\Component\PasswordHasher\Hasher\UserPasswordHasherInterface;
-use JMS\Serializer\SerializerInterface as JMSInterface;
 
 class UserController extends AbstractController
 {
-    protected $cache;
     protected $userRepository;
+    protected $cache;
     protected $serializer;
     protected $validator;
     protected $hasher;
 
     public function __construct(
-        CacheInterface $cache,
         UserRepository $userRepository,
-        SerializerInterface $serializer,
+        CacheInterface $cache,
+        JMSInterface $serializer,
         ValidatorInterface $validator,
         UserPasswordHasherInterface $hasher
-    ) {
-        $this->cache = $cache;
+    ) 
+    {
         $this->userRepository = $userRepository;
+        $this->cache = $cache;
         $this->serializer = $serializer;
         $this->validator = $validator;
         $this->hasher = $hasher;
@@ -48,12 +46,35 @@ class UserController extends AbstractController
 
     /**
      * @Route("/user", name="api_user_list", methods={"GET"})
+     * 
+     * @OA\Get(summary="Get list of BileMo users")
+     * @OA\Response(
+     *     response=Response::HTTP_OK,
+     *     description="Returns the list of users"
+     * )
+     * @OA\Response(
+     *     response=Response::HTTP_UNAUTHORIZED,
+     *     description="Invalid JWT Token"
+     * )
+     * @OA\Parameter(
+     *     name="page",
+     *     in="query",
+     *     description="The page number",
+     *     @OA\Schema(type="int", default = "1")
+     * )
+     * @OA\Parameter(
+     *     name="limit",
+     *     in="query",
+     *     description="Number of users by page",
+     *     @OA\Schema(type="int", default = 5)
+     * )
+     * @OA\Tag(name="Users")
      */
     public function index(
         PaginatorInterface $paginator,
-        Request $request,
-        JMSInterface $serializer
-    ): Response {
+        Request $request
+    ): Response 
+    {
         $userRepository = $this->userRepository;
 
         $customer_id = $this->getUser()->getCustomer()->getId();
@@ -80,7 +101,7 @@ class UserController extends AbstractController
             'meta' => $pagination->getPaginationData()
         ];
 
-        $json = $serializer->serialize(
+        $json = $this->serializer->serialize(
             $result,
             'json',
             SerializationContext::create()->setGroups(array('user:list'))
@@ -95,8 +116,23 @@ class UserController extends AbstractController
 
     /**
      * @Route("/user/{id}", name="api_user_details", methods={"GET"})
+     * 
+     * @OA\Get(summary="Get details of a user")
+     * @OA\Response(
+     *     response=Response::HTTP_OK,
+     *     description="Returns a user"
+     * )
+     * @OA\Response(
+     *     response=Response::HTTP_NOT_FOUND,
+     *     description="User not found"
+     * )
+     * @OA\Response(
+     *     response=Response::HTTP_UNAUTHORIZED,
+     *     description="Invalid JWT Token"
+     * )
+     * @OA\Tag(name="Users")
      */
-    public function show($id, JMSInterface $serializer)
+    public function show($id): Response
     {
         $customerId = $this->getUser()->getCustomer()->getId();
 
@@ -110,7 +146,7 @@ class UserController extends AbstractController
             throw new Exception('unauthorized', Response::HTTP_UNAUTHORIZED);
         }
 
-        $json = $serializer->serialize(
+        $json = $this->serializer->serialize(
             $userBdd,
             'json',
             SerializationContext::create()->setGroups(array('user:details'))
@@ -125,12 +161,53 @@ class UserController extends AbstractController
 
     /**
      * @Route("/user", name="api_user_create", methods={"POST"})
+     * 
+     * @OA\Post(summary="Create a new user")
+     * @OA\RequestBody(
+     *     description="Create a new user",
+     *     required=true,
+     *     @OA\MediaType(
+     *         mediaType="application/Json",
+     *         @OA\Schema(
+     *             type="object",
+     *             @OA\Property(
+     *                 property="fullName",
+     *                 description="Fisrtname and lastname for user identification",
+     *                 type="string"
+     *             ),     
+     *             @OA\Property(
+     *                 property="email",
+     *                 description="User's email",
+     *                 type="string"
+     *             ),
+     *             @OA\Property(
+     *                 property="password",
+     *                 description="Password chosen by the user with a minimum of height characters, one uppercase letter, one special character and one number",
+     *                 type="string"
+     *             )
+     *         )
+     *     )
+     * )
+     * @OA\Response(
+     *     response=Response::HTTP_CREATED,
+     *     description="Returns the new user"
+     * )
+     * @OA\Response(
+     *     response=Response::HTTP_BAD_REQUEST,
+     *     description="Bad Json syntax or incorrect data"
+     * )
+     * @OA\Response(
+     *     response=Response::HTTP_UNAUTHORIZED,
+     *     description="Invalid JWT Token"
+     * )
+     * @OA\Tag(name="Users")
      */
     public function create(
         Request $request,
         EntityManagerInterface $em,
         CustomerRepository $customerRepository
-    ) {
+    ): Response
+    {
         $user = $this->serializer->deserialize($request->getContent(), User::class, 'json');
 
         $customerId = $this->getUser()->getCustomer()->getId();
@@ -153,36 +230,74 @@ class UserController extends AbstractController
 
         $this->cache->delete('users_' . $customerId);
 
-        return $this->json(
+        $json = $this->serializer->serialize(
             $user,
+            'json',
+            SerializationContext::create()->setGroups(array('user:list'))
+        );
+
+        return new Response(
+            $json,
             Response::HTTP_CREATED,
-            [],
-            ['groups' => 'user:details']
+            array('Content-Type' => 'application/json')
         );
     }
 
     /**
      * @Route("/user/{id}", name="api_user_update", methods={"PUT"})
+     * 
+     * @OA\Put(summary="Update a user")
+     * @OA\RequestBody(
+     *     description="User data to modify : put only the fields to modify",
+     *     required=true,
+     *     @OA\MediaType(
+     *         mediaType="application/Json",
+     *         @OA\Schema(
+     *             type="object",
+     *             @OA\Property(
+     *                 property="fullName",
+     *                 description="Fisrtname and lastname for user identification",
+     *                 type="string"
+     *             ),     
+     *             @OA\Property(
+     *                 property="email",
+     *                 description="User's email",
+     *                 type="string"
+     *             ),
+     *             @OA\Property(
+     *                 property="password",
+     *                 description="Password chosen by the user",
+     *                 type="string"
+     *             )
+     *         )
+     *     )
+     * )
+     * @OA\Response(
+     *     response=Response::HTTP_OK,
+     *     description="Returns a modified user"
+     * )
+     * @OA\Response(
+     *     response=Response::HTTP_BAD_REQUEST,
+     *     description="Bad Json syntax or incorrect data"
+     * )
+     * @OA\Response(
+     *     response=Response::HTTP_UNAUTHORIZED,
+     *     description="Invalid JWT Token"
+     * )
+     * @OA\Tag(name="Users")
      */
-    public function update(Request $request, EntityManagerInterface $em, $id)
+    public function update(Request $request, EntityManagerInterface $em, $id): Response
     {
         $user = $this->userRepository->findOneBy(['id' => $id]);
 
         $userJson = $this->serializer->deserialize($request->getContent(), User::class, 'json');
-
-        $errors = $this->validator->validate($userJson);
-
-        if (count($errors) > 0) {
-            $errorsString = (string) $errors;
-            throw new Exception($errorsString, Response::HTTP_BAD_REQUEST);
-        }
 
         if ($userJson->getEmail()) {
             $user->setEmail($userJson->getEmail());
         }
 
         if ($userJson->getPassword()) {
-            $user->setPassword($this->hasher->hashPassword($user, $userJson->getPassword()));
+            $user->setPassword($userJson->getPassword());
         }
 
         if ($userJson->getfullName()) {
@@ -191,22 +306,43 @@ class UserController extends AbstractController
 
         $user->setUpdatedAt(new DateTime());
 
+        $errors = $this->validator->validate($user);
+
+        if (count($errors) > 0) {
+            $errorsString = (string) $errors;
+            throw new Exception($errorsString, Response::HTTP_BAD_REQUEST);
+        }
+
+        $user->setPassword($this->hasher->hashPassword($user, $userJson->getPassword()));
+
         $em->flush();
 
         $customerId = $this->getUser()->getCustomer()->getId();
 
         $this->cache->delete('users_' . $customerId);
 
-        return $this->json(
-            [],
-            Response::HTTP_OK
+        return new Response(
+            json_encode([]),
+            Response::HTTP_OK,
+            array('Content-Type' => 'application/json')
         );
     }
 
     /**
      * @Route("/user/{id}", name="api_user_delete", methods={"DELETE"})
+     * 
+     * @OA\Delete(summary="Delete a user")
+     * @OA\Response(
+     *     response=Response::HTTP_NO_CONTENT,
+     *     description="Delete a user"
+     * )
+     * @OA\Response(
+     *     response=Response::HTTP_UNAUTHORIZED,
+     *     description="Invalid JWT Token"
+     * )
+     * @OA\Tag(name="Users")
      */
-    public function delete(User $user, EntityManagerInterface $em)
+    public function delete(User $user, EntityManagerInterface $em): Response
     {
         $em->remove($user);
         $em->flush();
@@ -215,9 +351,10 @@ class UserController extends AbstractController
 
         $this->cache->delete('users_' . $customerId);
 
-        return $this->json(
-            [],
-            Response::HTTP_NO_CONTENT
+        return new Response(
+            json_encode([]),
+            Response::HTTP_NO_CONTENT,
+            array('Content-Type' => 'application/json')
         );
     }
 }
